+++ consensus/src/consensusdb/consensus_ghostdag.rs
@@ -208,7 +208,7 @@ pub struct DbGhostdagStore {
 }
 
 impl DbGhostdagStore {
-    pub fn new(db: Arc<DBStorage>, level: BlockLevel, cache_size: u64) -> Self {
+    pub fn new(db: Arc<DBStorage>, level: BlockLevel, cache_size: usize) -> Self {
         Self {
             db: Arc::clone(&db),
             level,
@@ -217,7 +217,7 @@ impl DbGhostdagStore {
         }
     }
 
-    pub fn clone_with_new_cache(&self, cache_size: u64) -> Self {
+    pub fn clone_with_new_cache(&self, cache_size: usize) -> Self {
         Self::new(Arc::clone(&self.db), self.level, cache_size)
     }
